Ciallo～(∠・ω< )⌒★

+++ Library_chacker/ds/Static_Range_Inversions_Query.cpp
@@ -1,11 +1,14 @@
 #include "MeIoN_Lib/Z_H/MeioN.hpp"
+#include "MeIoN_Lib/Z_H/FH.hpp"
 #include "MeIoN_Lib/MeIoN_all.hpp"
+#include "MeIoN_Lib/IO/fmt.hpp"
+#include "MeIoN_Lib/IO/fast_io.hpp"
 #include "MeIoN_Lib/ds/off_line/range_quis/range_inverse_count.hpp"
 
 // #define tests
 void Yorisou() {
   LL(n, q);
-  VEC(ll, a, n);
+  VEC(int, a, n);
   range_inverse_count seg(a);
   seg.read_quis<0>(q);
   for (ll x : seg.count()) {
